restructure pt 1
